little fix forbinding combobox to int

+++ lightbinding.h
@@ -238,8 +238,8 @@ inline QComboBox* bindCbFromVal(QObject* connector, TVal& val, QComboBox* cb = n
         }
     };
     connector->connect(cb, qOverload<int>(&QComboBox::currentIndexChanged), lambda);
-    if (cb->count() <= val) return cb;
-    cb->setCurrentIndex(val);
+    if (cb->count() <= static_cast<int>(val)) return cb;
+    cb->setCurrentIndex(static_cast<int>(val));
     return cb;
 }
 }
